Adding logstash config generator
