+++ cid/research.chop.edu/logstash/supervisor_logstash_run.conf
@@ -1,10 +1,7 @@
 [supervisord]
 nodaemon=false
-;logfile = /var/log/supervisor/supervisord.log
-;loglevel = debug
 
 [program:logstash_omop_harvest]
-priority=6
 directory=/opt/logstash
 startretries=3
 redirect_stderr = true
